requirements.txt
Flask
torch
torchvision
tensorflow
Pillow
numpy

